Refactored filter tabs in analytics UTM
ref https://linear.app/ghost/issue/PROD-2555/add-dropdown-component-to-statsposts-components

- The current `SourceTabs` component is very much restricted to be used only in the context of the Top Sources table in Analytics. In the future we want to use the same frontend logic in other places (e.g. Growth table) so we need to factor out the campaign dropdown and allow adding various tabs (not only "Sources" and "Campaigns"), in a more composable way.
