+++ apps/posts/src/views/PostAnalytics/Web/components/Sources.tsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import SourceIcon from '../../components/SourceIcon';
 import {BaseSourceData, ProcessedSourceData, extendSourcesWithPercentages, processSources, useTinybirdQuery} from '@tryghost/admin-x-framework';
-import {Button, CampaignType, Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle, DataList, DataListBar, DataListBody, DataListHead, DataListHeader, DataListItemContent, DataListItemValue, DataListItemValueAbs, DataListItemValuePerc, DataListRow, HTable, LucideIcon, Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger, SkeletonTable, SourceTabs, TabType, formatNumber, formatPercentage, formatQueryDate, getRangeDates} from '@tryghost/shade';
+import {Button, CampaignType, Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle, DataList, DataListBar, DataListBody, DataListHead, DataListHeader, DataListItemContent, DataListItemValue, DataListItemValueAbs, DataListItemValuePerc, DataListRow, HTable, LucideIcon, Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger, SkeletonTable, TabType, TableFilterCampaignTab, TableFilterTab, TableFilterTabs, formatNumber, formatPercentage, formatQueryDate, getRangeDates} from '@tryghost/shade';
 import {getAudienceQueryParam} from '../../components/AudienceSelect';
 import {getPeriodText} from '@src/utils/chart-helpers';
 import {useGlobalData} from '@src/providers/PostAnalyticsContext';
@@ -269,12 +269,15 @@ export const Sources: React.FC<SourcesCardProps> = ({
             <CardContent className='overflow-hidden'>
                 {utmTrackingEnabled && (
                     <div className='mb-4'>
-                        <SourceTabs
+                        <TableFilterTabs
                             selectedCampaign={selectedCampaign}
                             selectedTab={selectedTab}
                             onCampaignChange={setSelectedCampaign}
                             onTabChange={setSelectedTab}
-                        />
+                        >
+                            <TableFilterTab value='sources'>Sources</TableFilterTab>
+                            <TableFilterCampaignTab />
+                        </TableFilterTabs>
                     </div>
                 )}
                 <div className='h-[1px] w-full bg-border' />
