+++ apps/shade/src/index.ts
@@ -46,8 +46,8 @@ export * from './components/layout/view-header';
 
 // Feature components — Complete functional components (share modal, etc.)
 export {default as PostShareModal} from './components/features/post_share_modal';
-export {default as SourceTabs} from './components/features/sources/source-tabs';
-export type {CampaignType, TabType} from './components/features/sources/source-tabs';
+export * from './components/features/table-filter-tabs/table-filter-tabs';
+export type {CampaignType, TabType} from './components/features/table-filter-tabs/table-filter-tabs';
 
 // Third party components
 export * as Recharts from 'recharts';
